fix(services): switch L1 services to std and resolve all errors
Per the architecture plan, L1 services are allowed full std access.
Remove the incorrect #![no_std] constraint from all three service
crates and fix cascading compilation and test errors.

semantic_memory:
- Remove #![no_std] from lib.rs, enabling std types (Path, File, etc.)
- Wire orphaned modules (allocation, eviction, indexing, tiers)
- Fix broken use-block syntax in mem_ipc_handler.rs
- Remove extra closing brace in performance_baseline.rs
- Add thiserror::Error derive to MemoryError
- Add Ord derives to MemoryOperation for BTreeSet usage
- Add BitOr impl for MountFlags
- Fix borrow-after-move in context_sizing.rs via clone()
- Fix field names (used_bytes -> allocated_bytes)
- Fix mutability and borrow checker issues in tests
- Replace alloc:: with std:: in integration tests

gpu_accelerator:
- Remove #![no_std] from lib.rs
- Wire orphaned modules (profiling, scheduling, vram_management)
- Fix broken use-block in phase0_completion_report.rs
- Add Ord derives to FrameworkType for BTreeMap key usage
- Fix GpuDevice construction to use ::new() constructor
- Fix gen_val variable reference in scheduler_feedback tests
- Fix GpuDeviceID construction in error recovery tests
- Fix field names in integration tests
- Make RightSizer fields public for test access

tool_registry_telemetry:
- Remove #![no_std] from lib.rs
- Wire orphaned modules (registry, telemetry, journal, compliance_l1)
- Add thiserror::Error derive to ToolError
- Add Ord derives to ToolID and AgentID
- Add CapID type with from_bytes/as_bytes/Display
- Fix timestamp_ms -> timestamp_ns (3 occurrences)
- Fix Error -> ToolError naming across persistent_event_logger,
  retention_policy, and performance_baselines
- Disambiguate schema_version() calls in cef_format tests
- Fix TypeSchema construction in integration tests
- Replace alloc:: with String::from() in integration tests

+++ services/gpu_accelerator/tests/l1_scaffold_tests.rs
@@ -221,7 +221,7 @@ fn test_profiling_workflow() {
     // Simulate kernel execution and profiling
     for kernel_id in 0..5 {
         // Schedule
-        let _ = scheduler.schedule_kernel(kernel_id, 20 + kernel_id * 5);
+        let _ = scheduler.schedule_kernel(kernel_id, (20 + kernel_id * 5) as u32);
 
         // Profile execution
         let timing = GpuMs::new(50, 20, 10);
